# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: API documentation https://tracking.complexus.uk.
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.15.3'
servers:
  - url: https://tracking.complexus.uk/api
    description: Production Server
  - url: https://dev.complexus.uk/api
    description: Devleopment Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /geofences:
    get:
      operationId: getGeofences
      summary: Fetch a list of Geofences
      description: Without params, it returns a list of Geofences the user has access to
      tags:
        - Geofences
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
        - name: deviceId
          in: query
          description: >-
            Standard users can use this only with _deviceId_s, they have access
            to
          schema:
            type: integer
        - name: groupId
          in: query
          description: >-
            Standard users can use this only with _groupId_s, they have access
            to
          schema:
            type: integer
        - name: refresh
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          description: Limit the number of returned results
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
        - name: keyword
          in: query
          description: Search keyword filter
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
        '401':
          description: Unauthorized
          content: {}
    post:
      operationId: postGeofences
      summary: Create a Geofence
      tags:
        - Geofences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Geofence'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
...
  /reports/combined:
    get:
      operationId: getReportsCombined
      summary: Fetch a combined route, Events and Positions report for the Devices or Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CombinedReportItem'
        '401':
          description: Unauthorized
          content: {}
  /reports/route:
    get:
      operationId: getReportsRoute
      summary: >-
        Fetch a list of Positions within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
        '401':
          description: Unauthorized
          content: {}
...
  /reports/events:
    get:
      operationId: getReportsEvents
      summary: Fetch a list of Events within the time period for the Devices or Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: type
          in: query
          description: '% can be used to return events of all types'
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
          content: {}
...
  /reports/geofences:
    get:
      operationId: getReportsGeofences
      summary: >-
        Fetch geofence enter/exit intervals within the time period for Devices
        or Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: geofenceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportGeofences'
        '401':
          description: Unauthorized
          content: {}
  /reports/summary:
    get:
      operationId: getReportsSummary
      summary: >-
        Fetch a list of ReportSummary within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportSummary'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportSummary'
        '401':
          description: Unauthorized
          content: {}
...
  /reports/trips:
    get:
      operationId: getReportsTrips
      summary: >-
        Fetch a list of ReportTrips within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTrips'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTrips'
        '401':
          description: Unauthorized
          content: {}
...
  /reports/stops:
    get:
      operationId: getReportsStops
      summary: >-
        Fetch a list of ReportStops within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportStops'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportStops'
        '401':
          description: Unauthorized
          content: {}
...
components:
  schemas:
    Position:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique position record identifier
        deviceId:
          type: integer
          format: int64
          description: Identifier of the device that reported this position
        protocol:
          type: string
          description: Device protocol name that produced the message
        deviceTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        fixTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        serverTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        valid:
          type: boolean
          description: Indicates whether the position was marked as valid by the device
        latitude:
          type: number
          description: Latitude in decimal degrees
        longitude:
          type: number
          description: Longitude in decimal degrees
        altitude:
          type: number
          description: Altitude above sea level in meters
        speed:
          type: number
          description: in knots
        course:
          type: number
          description: Heading in degrees (0-360) where 0 is true north
        address:
          type: string
          description: Resolved reverse-geocoded address if available
        accuracy:
          type: number
          description: Estimated positional accuracy in meters when provided
        network:
          type: object
          description: Network metadata (e.g. cell or Wi‑Fi data) supplied by the device
          properties: {}
        geofenceIds:
          type: array
          items:
            type: integer
          description: List of geofence ids applicable to this position
        attributes:
          type: object
          description: Custom key-value attributes sent by the device or enrichments
          properties: {}
...
    Geofence:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the geofence
        name:
          type: string
          description: Human-readable name shown in lists and maps
        description:
          type: string
          description: Details about the geofence for display in the UI
        area:
          type: string
          description: Geofence area definition encoded as a WKT string
        calendarId:
          type: integer
          format: int64
          description: Calendar identifier limiting when the geofence is active
        attributes:
          type: object
          properties: {}
          description: Custom key-value pairs for integrations or UI overrides
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique event identifier
        type:
          type: string
          description: Event type name
        eventTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        deviceId:
          type: integer
          format: int64
          description: Device associated with the event
        positionId:
          type: integer
          format: int64
          description: Position record linked to the event when applicable
        geofenceId:
          type: integer
          format: int64
          description: Geofence referenced by the event if any
        maintenanceId:
          type: integer
          description: Maintenance record tied to the event
        attributes:
          type: object
          properties: {}
          description: Additional event-specific attributes
    CombinedReportItem:
      type: object
      properties:
        deviceId:
          type: integer
          format: int64
          description: Device identifier for the report item
        route:
          type: array
          description: Simplified route as `[longitude, latitude]` pairs
          items:
            type: array
            items:
              type: number
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer